+++ R/shortcuts.R
@@ -1,4 +1,4 @@
-# Copyright 2017 Province of British Columbia
+# Copyright 2018 Province of British Columbia
   #
   # Licensed under the Apache License, Version 2.0 (the "License");
   # you may not use this file except in compliance with the License.
@@ -25,6 +25,8 @@
 #'
 #' @details Type `?bcmaps.rdata::airzones` for details.
 #'
+#' 
+#'
 #' @examples
 #' \dontrun{
 #' my_layer <- airzones()
@@ -45,6 +47,8 @@ airzones <- function(class = 'sf') {
 #' @return The spatial layer of `bc_bound` in the desired class
 #'
 #' @details Type `?bcmaps.rdata::bc_bound` for details.
+#'
+#' 
 #'
 #' @examples
 #' \dontrun{
@@ -67,6 +71,8 @@ bc_bound <- function(class = 'sf') {
 #'
 #' @details Type `?bcmaps.rdata::bc_bound_hres` for details.
 #'
+#' 
+#'
 #' @examples
 #' \dontrun{
 #' my_layer <- bc_bound_hres()
@@ -78,6 +84,29 @@ bc_bound_hres <- function(class = 'sf') {
    get_layer('bc_bound_hres', class = class)
 }
 
+#' BC Major Cities Points 1:2,000,000 (Digital Baseline Mapping)
+#'
+#' You must have the `bcmaps.rdata` package installed to use this function.
+#'
+#' @param class what class you want the object in? `"sf"` (default) or `"sp"`.
+#'
+#' @return The spatial layer of `bc_cities` in the desired class
+#'
+#' @details Type `?bcmaps.rdata::bc_cities` for details.
+#'
+#' 
+#'
+#' @examples
+#' \dontrun{
+#' my_layer <- bc_cities()
+#' my_layer_sp <- bc_cities(class = 'sp')
+#' }
+#'
+#' @export
+bc_cities <- function(class = 'sf') {
+   get_layer('bc_cities', class = class)
+}
+
 #' British Columbia Ecoprovinces
 #'
 #' You must have the `bcmaps.rdata` package installed to use this function.
@@ -87,6 +116,8 @@ bc_bound_hres <- function(class = 'sf') {
 #' @return The spatial layer of `ecoprovinces` in the desired class
 #'
 #' @details Type `?bcmaps.rdata::ecoprovinces` for details.
+#'
+#' 
 #'
 #' @examples
 #' \dontrun{
@@ -109,6 +140,8 @@ ecoprovinces <- function(class = 'sf') {
 #'
 #' @details Type `?bcmaps.rdata::ecoregions` for details.
 #'
+#' 
+#'
 #' @examples
 #' \dontrun{
 #' my_layer <- ecoregions()
@@ -129,6 +162,8 @@ ecoregions <- function(class = 'sf') {
 #' @return The spatial layer of `ecosections` in the desired class
 #'
 #' @details Type `?bcmaps.rdata::ecosections` for details.
+#'
+#' 
 #'
 #' @examples
 #' \dontrun{
@@ -151,6 +186,8 @@ ecosections <- function(class = 'sf') {
 #'
 #' @details Type `?bcmaps.rdata::gw_aquifers` for details.
 #'
+#' 
+#'
 #' @examples
 #' \dontrun{
 #' my_layer <- gw_aquifers()
@@ -171,6 +208,8 @@ gw_aquifers <- function(class = 'sf') {
 #' @return The spatial layer of `hydrozones` in the desired class
 #'
 #' @details Type `?bcmaps.rdata::hydrozones` for details.
+#'
+#' 
 #'
 #' @examples
 #' \dontrun{
@@ -193,6 +232,8 @@ hydrozones <- function(class = 'sf') {
 #'
 #' @details Type `?bcmaps.rdata::municipalities` for details.
 #'
+#' @seealso [combine_nr_rd()] to combine Regional Districts and the Northern Rockies Regional Municipality into one layer
+#'
 #' @examples
 #' \dontrun{
 #' my_layer <- municipalities()
@@ -213,6 +254,8 @@ municipalities <- function(class = 'sf') {
 #' @return The spatial layer of `nr_areas` in the desired class
 #'
 #' @details Type `?bcmaps.rdata::nr_areas` for details.
+#'
+#' 
 #'
 #' @examples
 #' \dontrun{
@@ -235,6 +278,8 @@ nr_areas <- function(class = 'sf') {
 #'
 #' @details Type `?bcmaps.rdata::nr_districts` for details.
 #'
+#' 
+#'
 #' @examples
 #' \dontrun{
 #' my_layer <- nr_districts()
@@ -255,6 +300,8 @@ nr_districts <- function(class = 'sf') {
 #' @return The spatial layer of `nr_regions` in the desired class
 #'
 #' @details Type `?bcmaps.rdata::nr_regions` for details.
+#'
+#' 
 #'
 #' @examples
 #' \dontrun{
@@ -277,6 +324,8 @@ nr_regions <- function(class = 'sf') {
 #'
 #' @details Type `?bcmaps.rdata::regional_districts` for details.
 #'
+#' @seealso [combine_nr_rd()] to combine Regional Districts and the Northern Rockies Regional Municipality into one layer
+#'
 #' @examples
 #' \dontrun{
 #' my_layer <- regional_districts()
@@ -297,6 +346,8 @@ regional_districts <- function(class = 'sf') {
 #' @return The spatial layer of `water_districts` in the desired class
 #'
 #' @details Type `?bcmaps.rdata::water_districts` for details.
+#'
+#' 
 #'
 #' @examples
 #' \dontrun{
@@ -319,6 +370,8 @@ water_districts <- function(class = 'sf') {
 #'
 #' @details Type `?bcmaps.rdata::water_precincts` for details.
 #'
+#' 
+#'
 #' @examples
 #' \dontrun{
 #' my_layer <- water_precincts()
@@ -339,6 +392,8 @@ water_precincts <- function(class = 'sf') {
 #' @return The spatial layer of `watercourses_15M` in the desired class
 #'
 #' @details Type `?bcmaps.rdata::watercourses_15M` for details.
+#'
+#' 
 #'
 #' @examples
 #' \dontrun{
@@ -361,6 +416,8 @@ watercourses_15M <- function(class = 'sf') {
 #'
 #' @details Type `?bcmaps.rdata::watercourses_5M` for details.
 #'
+#' 
+#'
 #' @examples
 #' \dontrun{
 #' my_layer <- watercourses_5M()
@@ -381,6 +438,8 @@ watercourses_5M <- function(class = 'sf') {
 #' @return The spatial layer of `wsc_drainages` in the desired class
 #'
 #' @details Type `?bcmaps.rdata::wsc_drainages` for details.
+#'
+#' 
 #'
 #' @examples
 #' \dontrun{